[FIX] base_name_search_improved: Manage limit None
When we are searching within a m2m result that will call the respective
name_search() but will be build on the ovs.expression.parse() method with
limit=None. This is making our name_search since we are trying this as int
value.

Here is the code when our limit is flushed and set to None

> /odoo/osv/expression.py(956)parse()
-> ids2 = [x[0] for x in comodel.name_search(right, domain or [], op2, limit=None)]

https://github.com/odoo/odoo/blob/11.0/odoo/osv/expression.py?utf8=%E2%9C%93#L956